Accounts and Users have been added as a functionality
An account can now be created that should keep anything to do with the
financial side aside for a User. That is there is a hashmap of users
that have a bank account and their username and password. Tests have
been written for it and this should account for both new users and
testing if existing ones can login.