Bank can transfer money from sender to recipient
The bank class now can use transferFunds(sender, recipient, amount) to
send money from a sender to a recipient with a specified amount. It
verifies the sender, recipient, and the amount are all valid within the
context of the bank. Then it verifies if the amount is sufficient based
on the current balance of the sender. If all is valid, then the amount
is withdrawn from the sender and deposited in the recipients bank
account.