Finished integrating admin functionality to CLI interface
An admin can now view all users in the bank, all the money the bank has,
and add new admin users through the CLI.
An admin has a separate menu interface than the user does (for now).
Tests have been created for this CLI integration and all functionality
from previous tasks remains intact.
The way to access admin stuff is to sign in through the initialized
"Ownser" of the bank:
--> Username: "owner"
--> Password: "verysecurePassword43"

In deployment that would not be available obviously

P.S.: Right now, this is just to expose some stuff an admin can do. They
can't really deposit, withdraw, etc. That may require the program to
somehow attach a virtual child panel to the admin to make it think it is
a regular user and such. However for now, this is just exposing that an
admin can actual do stuff in the CLI.